Break various asset functions out into separate files, including a new one for updateAssets.
